requirements.txt
flask
pymongo[srv]
python-dotenv
gunicorn
bcrypt
email.mime.text
email.mime.multipart